requirements.txt

requests
PyYAML
